rename undefined category to 'All Products' + refactor
