+++ src/components/Products/Products.module.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/sass-functions/functions';
+@import '../../assets/sass-functions/functions';
 
 main {
   flex-grow: 1;
