+++ src/components/Products/Sidebar/PriceRange.tsx
