+++ src/components/Products/Sidebar/filterOptions/getCheckboxOptions.ts
